Change `bin/rails` to `rails`
From reading the Rails code, it appears that the global `rails` command
actually just calls `bin/rails` if it exists. Given this, calling
`bin/rails` is just using more characters for the same result.

I'm not sure why I've had issues with it in the past but it's entirely
possible that it was user error and, because I assumed that Rails was in
the wrong, I failed to properly investigate what was going on